requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   groq
    #   httpx
beautifulsoup4==4.13.3
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   pip-tools
    #   surya-ocr
colorama==0.4.6
    # via
    #   build
    #   click
    #   loguru
    #   tqdm
customtkinter==5.2.2
    # via -r requirements.in
darkdetect==0.8.0
    # via customtkinter
distro==1.9.0
    # via groq
exceptiongroup==1.2.2
    # via anyio
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
filetype==1.2.0
    # via surya-ocr
fire==0.7.0
    # via manga-ocr
fsspec==2025.3.0
    # via
    #   huggingface-hub
    #   torch
fugashi==1.4.0
    # via manga-ocr
groq==0.20.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via groq
huggingface-hub==0.29.3
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jaconv==0.4.0
    # via manga-ocr
jinja2==3.1.6
    # via torch
keyboard==0.13.5
    # via -r requirements.in
loguru==0.7.3
    # via manga-ocr
manga-ocr==0.1.14
    # via -r requirements.in
markdown==3.7
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mouseinfo==0.1.3
    # via pyautogui
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   manga-ocr
    #   opencv-python-headless
    #   transformers
opencv-python-headless==4.11.0.86
    # via surya-ocr
packaging==24.2
    # via
    #   build
    #   customtkinter
    #   huggingface-hub
    #   transformers
pillow==10.4.0
    # via
    #   manga-ocr
    #   pyscreeze
    #   surya-ocr
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.7
    # via surya-ocr
pyautogui==0.9.54
    # via -r requirements.in
pydantic==2.11.1
    # via
    #   groq
    #   pydantic-settings
    #   surya-ocr
pydantic-core==2.33.0
    # via pydantic
pydantic-settings==2.8.1
    # via surya-ocr
pygetwindow==0.0.9
    # via pyautogui
pymsgbox==1.0.9
    # via pyautogui
pypdfium2==4.30.0
    # via surya-ocr
pyperclip==1.9.0
    # via
    #   manga-ocr
    #   mouseinfo
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
python-dotenv==1.1.0
    # via
    #   pydantic-settings
    #   surya-ocr
pytweening==1.2.0
    # via pyautogui
pywin32==310
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via transformers
sniffio==1.3.1
    # via
    #   anyio
    #   groq
soupsieve==2.6
    # via beautifulsoup4
surya-ocr==0.13.1
    # via -r requirements.in
sympy==1.13.1
    # via torch
termcolor==2.5.0
    # via fire
tokenizers==0.21.1
    # via transformers
tomli==2.2.1
    # via
    #   build
    #   pip-tools
torch==2.6.0
    # via
    #   manga-ocr
    #   surya-ocr
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.50.3
    # via
    #   manga-ocr
    #   surya-ocr
typing-extensions==4.13.0
    # via
    #   anyio
    #   beautifulsoup4
    #   groq
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
unidic-lite==1.0.8
    # via manga-ocr
urllib3==2.3.0
    # via requests
wheel==0.45.1
    # via pip-tools
win32-setctime==1.2.0
    # via loguru

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
